QEMU/logging: send bits logs to a pre-determined IO port instead of a file
Sometimes it is necessary to send bits test logs outside of the running system
as it helps in offline analysis of test failures or when run in an automated
test framework, fail the running test. When bits is run from within QEMU
environment, isa-debugcon device can be used to send the test logs to the test
framework. isa-debugcon is generally used by bioses (edk2 and seabios) to write
out the debug logs to the IO port 0x402 which are then used to analyze and
debug bioses. Following QEMU commandline can be used for example to enable
debugcon:

./qemu-system-x86_64 -chardev file,path=/tmp/debugcon.txt,id=debugcon \
  -device isa-debugcon,iobase=0x403,chardev=debugcon

More details can be seen here: https://www.seabios.org/Debugging

This change adds a new feature in bits where the logs are written to the IO
port 0x403. This port is reserved explicitly for bits logs and not used by
bioses to send their own logs.

Currently it is not possible to change the IO port address and the QEMU
commandline used to configure isa-debugcon must specify the same IO port
address as is being used by bits. A future change might make it possible to
specify/change the IO port address.

Signed-off-by: Ani Sinha <ani@anisinha.ca>